VIDEO-RENDERER-GCP/requirements.txt
annotated-types==0.7.0
av==13.1.0
azure-cognitiveservices-speech==1.45.0
azure-core==1.35.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
cloup==3.0.7
colorama==0.4.6
decorator==5.2.1
fastapi==0.104.1
ffmpeg==1.4
firebase-admin==6.2.0
glcontext==3.0.0
google-cloud-storage==2.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isosurfaces==0.1.2
manim==0.19.0
manim-voiceover
manim-voiceover-fixed==0.3.10
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
mutagen==1.47.0
networkx==3.5
numpy==2.3.2
pillow==11.3.0
pycairo==1.28.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.2
python-dotenv==0.21.0
python-multipart==0.0.6
python-slugify==8.0.4
pytest
requests==2.32.4
rich==14.1.0
scipy==1.16.1
screeninfo==0.8.1
setuptools==80.9.0
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
sox==1.5.0
srt==3.5.3
svgelements==1.9.6
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn[standard]==0.24.0
watchdog==6.0.0
# gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.21.6